+++ examples/scripts/nemo_gym/config.yaml
@@ -2,8 +2,13 @@
 model_name: "Qwen/Qwen3-4B-Instruct-2507"
 
 # Data
-dataset_path: "/home/ubuntu/Gym/resources_servers/workplace_assistant/data/train.jsonl"
-eval_dataset_path: "/home/ubuntu/Gym/resources_servers/workplace_assistant/data/validation.jsonl"
+dataset_path: "/path/to/Gym/resources_servers/workplace_assistant/data/train.jsonl"
+eval_dataset_path: "/path/to/Gym/resources_servers/workplace_assistant/data/validation.jsonl"
+
+# NeMo Gym server configs
+gym_configs:
+  - resources_servers/workplace_assistant/configs/workplace_assistant.yaml
+  - responses_api_models/vllm_model/configs/vllm_model_for_training.yaml
 
 # Logging
 output_dir: "outputs/nemo_gym"
@@ -27,12 +32,11 @@ optim: "adamw_torch_fused"
 weight_decay: 0.01
 vllm_importance_sampling_correction: true
 
-# Inference sampling parameters
+# Inference
 temperature: 1.0
 top_p: 0.999
 
-# Checkpointing and Eval
+# Checkpointing and eval
 save_steps: 10
 eval_strategy: "steps"
 eval_steps: 10
-
